src/authorizer/requirements.txt
boto3
python-jose
